requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# analysis dependencies
pandas>=1.0.4
scipy>=1.4.1
matplotlib>=3.2.1
numpy>=1.18.5
jupyter>=1.0.0
watermark>=2.0.2
sqlalchemy>=1.3.17
statsmodels>=0.11.1
seaborn>=0.10.1
scikit-learn>=0.23.1
factor_analyzer>=0.3.2
radon>=4.1.0
natsort